Code quality: move method to a specific file
